fix transmission-remote authentication
Some transmission-remote options fail if the --auth or --authenv options
are not given in earlier order on the command line. This patch fixes
this by setting --authenv before any other options.

Additionally, --authenv is used and TR_AUTH is set even when reading the
password from the config. This is better because if the password is
given as a command line argument, it is visible by any user on the
system. (For example in the output of "ps aux").

Closes #18