fix 'add' so that we send correct data to daemon (#20)
We were converting a byte object to str with str() but now we are
properly using bytes.decode().
This was a victim of the 2to3 automatic conversion.

This  is probably a common issue with python2->3 conversions. If
you're not sure what the big deal is just do
print(str(some_bytes_object)). Then compare it to
print(some_bytes_object.decode()). It should be obvious at that point.

Fixes #6