fix addch ERR crash when starting with no torrents
There was a bug introduced by commit e06d08d:
    scale_bytes: Simplify this function

tremc would crash when started against a transmission-daemon with no
torrents.

This was because scale_bytes(0) used to return 0K but now it returns
0.0K. The expected width of 2 was hardcoded. The new width of 4 causes
addch() to return ERR because it tries to draw outside the window.

Hardcoding the new width to 4 would resolve this issue. Instead,
this patch dynamically sets the width returned by scale_bytes().
This should make the code a tiny bit more readable and maybe avoid this
issue in the future.

There is one more magic number involved here which I don't see a good
way of getting rid of. So I made an illustrative comment.

Resolves #15