correct progress bars for unlimited ratio
Correct progress bars for unlimited ratio on individual torrents.

Without this patch:

*  Specify a global ratio. Say, 10.
*  Begin seeding a torrent that has ratio less than 10
*  Specify a seed ratio limit on that torrent of -1
*  Observe that the progress bar indicates the torrent is not done seeding

With this patch, individual torrents with unlimited ratio (-1) display
as 100% seeded. This is consistent with the behavior when global seed
ratio limit is disabled, where all torrents are displayed as 100%.

I've also rewritten the whole 'if' statement to be a bit more readable.

Resolves #5