fix alignment issue in torrent and file list
Commit e06d08d:
    scale_bytes: Simplify this function

introduced an issue where alignment of some fields broke in the
torrent and file list. This was because the maximum length of the values
returned by scale_bytes went from 5 to 8. For example:

    1000K => 1,000.0K

However, the alignment in both the torrent and file list was broken by
only one character. With this patch, the max is reduced by one
character. Just enough to properly align things again. From the
aforementioned example:

    1000.0K

Maybe some day we'll have a more consistent understanding and explicitly
specified rules about widths and alignments.

Resolves #31