add scroll inplace commands and bind to C-f/C-b
before this C-u did the same thing as C-b and C-d did the same thing
as C-f. This was a waste of a good keybinding.

Now C-b and C-f both scroll around the focused entry, while not
changing the focused entry. This has the affect of letting you see a
few torrents before what tremc is currently showing you, and a few
entries beyond what tremc is showing you.