Add support for profiles. A profile is defined in [Profiles] section of the config file:
profileNAME = FILTER#=ARG#=SORT

where NAME is the profile name,
FILTER is the filter to use (prefix with : for inverse),
ARG is the parameter for the filter (regex and tracker need one)
SORT is the sort order

A profile can be selected by a command line argument.
Profiles named 0..9 can also be selected at by pressing the appropriate number key.