fix bug where queuePosition sort in config is ignored
Queue was implemented in rpc version 14. The code checked for rpc
version >= 14 and tried to then enable queuePosition sort but this
happened after the config was parsed. Checking before the config is
parsed is not possible because the connection to the server is not
established that early.

This fix removes the check entirely. This means that for users who use
queuePosition sort in their configs and then downgrade for some reason
to an older transmission-daemon version, tremc will likely crash with an
exception.

This will not break any existing configs of old installations and for
this reason I consider it an acceptable fix, in the interest of time and
less code complexity.

fixes #64