Workaround for exit code
Python exits by SyetemExit exception, so catching using try...finally to make sure screen mode is restored at exit catches the exit and loses the exit code.
This is a workaround to save the exit code from the first (intentional) call to exit_prog for the second call in the finally clause to use.

This should solve https://github.com/tremc/tremc/issues/74