Remove queuePosition from sort options if server has no queue
This completes https://github.com/tremc/tremc/commit/61dae8dfb9ef88376e154a0ca1122ddee1321946 .

By default, queuePosition is in sort_options so that it may be used in the config file, which
is parsed before connecting to the server. If after connecting we see the server does not support
this feature, we remove it from the options and reset the sort order to the default sort order
if it was set in the config file to queuePosition.