Add `%t` optional parameter to file view command
`%t` will be replaced by four parameters on the command line:
1. server host name (from the config file)
2. download dir
3. incomplete dir
4. file name

If `%t` is on the command line, tremc does not test for the existence of the file.

This should allow for previewing files on a remote server (but requires a script to be written and knowledge of how to access the remote file system).

See https://github.com/tremc/tremc/issues/132